yeah this time is legit finished the ft_printf project

+++ ft_printf.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_printf.h                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tzi-qi <tzi-qi@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/08/20 16:29:54 by tzi-qi            #+#    #+#             */
+/*   Updated: 2022/08/20 16:30:05 by tzi-qi           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
@@ -11,10 +23,10 @@ int	ft_putchar(int c);
 int	ft_putstr(char *str);
 int	ft_putnbr(int n);
 int	ft_putnbr_u(unsigned int n);
-int	hexdecimal_l(int n);
-int	hexdecimal_u(int n);
-int	printf_pointer(int p);
-int printf_specifiers(va_list args, char format, int a);
-int ft_printf(const char *format, ...);
+int	hexdecimal_l(unsigned long long n);
+int	hexdecimal_u(unsigned long long n);
+int	printf_pointer(unsigned long long p);
+int	printf_specifiers(va_list args, char format, int a);
+int	ft_printf(const char *format, ...);
 
 #endif
